Creación de Web Security y sesión para coordinador

+++ GTICS_TA/src/main/java/com/example/gtics_ta/repository/AsistenciaRepository.java
@@ -3,6 +3,6 @@ package com.example.gtics_ta.repository;
 import com.example.gtics_ta.Entity.Asistencia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AsistenciaRepository extends JpaRepository<Asistencia, Long> {
+public interface AsistenciaRepository extends JpaRepository<Asistencia, Integer> {
 
 }
